+++ AirCasting/CreateSessionViews/ChooseSessionType/ChooseSessionTypeViewModel.swift
@@ -2,7 +2,6 @@
 //
 
 import Foundation
-import CoreBluetooth
 import Resolver
 
 enum ProceedToView {
