+++ AirCasting/Utils/Bluetooth/BluetoothConnectionObserver.swift
@@ -1,0 +1,8 @@
+// Created by Lunar on 16/11/2022.
+//
+
+import Foundation
+
+protocol BluetoothConnectionObserver: AnyObject {
+    func didDisconnect(device: NewBluetoothManager.BluetoothDevice)
+}
